refactor: 게시글 상세 데이터 가져오는 훅 부분 분리 및 적용
